services/matcher/requirements.txt
# Core ML & Data
torch>=2.0.0
transformers[torch]
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.21.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.0.2

# Web Service
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic-settings>=2.0.0

# Utilities
loguru>=0.7.0
PyYAML>=6.0
tqdm>=4.64.0
cryptography
datasets